+++ CHANGELOG.md
@@ -3,8 +3,6 @@ Changelog
 
 ##### Version 4.5.23 (WIP)
  * fix minor bugs
- * add ContactMessageAnswerType flash message translation
- * add weekday name translations in ExportCalendarToListBuilderPdf
 
 ##### Version 4.5.22 (2022-08-31)
  * fix bad green color definition in drawTeacherColorSpan Twig extension
@@ -14,6 +12,9 @@ Changelog
  * improve ContactMessageAnswerType
  * improve price column render in Tariff list
  * remove some hardcoded Admin list 'decimal' types
+ * add ContactMessageAnswerType flash message translation
+ * add weekday name translations in ExportCalendarToListBuilderPdf
+ * fix missing FRST & RCUR sepa batch actions
 
 ##### Version 4.5.21 (2021-11-26)
  * avoid SEPA receipts generation for empty units receipt case